Fix Gutenberg detection
Hi and thank you for this plugin. This PR fixes the detection of Gutenberg according to https://wordpress.stackexchange.com/questions/320653/how-to-detect-the-usage-of-gutenberg – please note that this is only a partial fix for the Gutenberg compatibility, since this way, the shortcodes are in text form in the Gutenberg editor. But, at least, the editor does not freeze by default as was the case for me before this change.